:bookmark_tabs: :: (#163) <XResponse> 응답 객체 계층 부여
