+++ src/main/java/app/jg/og/zamong/service/dream/sell/SellDreamService.java
@@ -2,7 +2,7 @@ package app.jg.og.zamong.service.dream.sell;
 
 import app.jg.og.zamong.dto.request.dream.selldream.SellDreamCostRequest;
 import app.jg.og.zamong.dto.request.dream.selldream.SellDreamRequest;
-import app.jg.og.zamong.dto.response.CreateDreamResponse;
+import app.jg.og.zamong.dto.response.dream.CreateDreamResponse;
 import app.jg.og.zamong.dto.response.dream.selldream.DoSellRequestDreamResponse;
 
 public interface SellDreamService {
