+++ src/main/java/app/jg/og/zamong/service/user/UserServiceImpl.java
@@ -2,7 +2,7 @@ package app.jg.og.zamong.service.user;
 
 import app.jg.og.zamong.dto.request.ChangePasswordRequest;
 import app.jg.og.zamong.dto.request.CheckIdDuplicationRequest;
-import app.jg.og.zamong.dto.response.UserInformationResponse;
+import app.jg.og.zamong.dto.response.user.UserInformationResponse;
 import app.jg.og.zamong.entity.dream.sharedream.ShareDreamRepository;
 import app.jg.og.zamong.entity.user.User;
 import app.jg.og.zamong.entity.user.UserRepository;
