+++ src/main/java/app/jg/og/zamong/service/user/auth/signup/UserSignUpService.java
@@ -2,7 +2,7 @@ package app.jg.og.zamong.service.user.auth.signup;
 
 import app.jg.og.zamong.dto.request.*;
 import app.jg.og.zamong.dto.response.StringResponse;
-import app.jg.og.zamong.dto.response.SignUpUserResponse;
+import app.jg.og.zamong.dto.response.user.SignUpUserResponse;
 
 public interface UserSignUpService {
