+++ src/test/java/app/jg/og/zamong/service/dream/ShareDreamServiceTest.java
@@ -1,7 +1,7 @@
 package app.jg.og.zamong.service.dream;
 
 import app.jg.og.zamong.dto.request.dream.sharedream.*;
-import app.jg.og.zamong.dto.response.CreateDreamResponse;
+import app.jg.og.zamong.dto.response.dream.CreateDreamResponse;
 import app.jg.og.zamong.dto.response.dream.sharedream.DoShareDreamResponse;
 import app.jg.og.zamong.entity.dream.attachment.AttachmentImage;
 import app.jg.og.zamong.entity.dream.attachment.AttachmentImageRepository;
